Fix untyped React.Components throws Error.
When a class was extending React.Component and did not specify
Prop/State types the code would fail with the error:
"PropertyDeclaration was unexpected"

Turns out class members are not supposed to be modified directly.
Using ts.updateClassDeclaration() seems to fix the problem and
also make other tests code look better.